Move JSON, install, uninstall, and update logic to its own class.
